% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_email_valido.R
\name{testar_validade_email}
\alias{testar_validade_email}
\title{Teste com expressão regular se o endereço de email tem uma estrutura válida}
\usage{
testar_validade_email(email)
}
\arguments{
\item{email}{Endereço de email a ser testado}
}
\value{
Valor lógico (TRUE/FALSE)
}
\description{
Teste com expressão regular se o endereço de email tem uma estrutura válida
}
\examples{
testar_validade_email("milz.bea@gmail.com")
}
