% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_email_valido.R
\name{teste_email_valido}
\alias{teste_email_valido}
\title{Teste com expressão regular se o endereço de email tem uma estrutura válida}
\usage{
teste_email_valido(email)
}
\arguments{
\item{email}{Endereço de email a ser testado}
}
\value{
Valor lógico (TRUE/FALSE)
}
\description{
Teste com expressão regular se o endereço de email tem uma estrutura válida
}
\examples{
teste_email_valido("milz.bea@gmail.com")
}
